Create FormManager, fix bag
1. Create FormManager  #689
2. fix Open only "Pipe" tab from menu "Settings" #1242
