+++ src/PrizmMainProject/Forms/MainChildForm/PrizmApplicationXtraForm.cs
@@ -62,128 +62,7 @@ namespace Prizm.Main.Forms.MainChildForm
             InitializeComponent();
             Bitmap bmp = Resources.prizma_appIcon_32;
             this.Icon = Icon.FromHandle(bmp.GetHicon());
-        }
-
-
-        /// <summary>
-        /// Gives the ChildForm index in corresponding childForms list that contains the element defined by id.
-        /// </summary>
-        /// <param name="formTypeName">string representation of form type name, for example</param>
-        /// <param name="id"></param>
-        /// <returns>The zero-based index of the first occurrence of an element that matches the
-        /// conditions defined by match, if found; otherwise, –1.</returns>
-        private int GetFormIndex(string formTypeName, Guid id)
-        {
-            int index = 0;
-
-            if (id != Guid.Empty)
-            {
-                index = childForms[formTypeName]
-                    .FindIndex(x => x is INewEditEntityForm && ((INewEditEntityForm)x).IsMatchedByGuid(id));
-            }
-            else
-            {
-                bool isSingleForm =
-                    formTypeName == typeof(SettingsXtraForm).Name ||
-                    formTypeName == typeof(NotificationXtraForm).Name;
-
-                if (!isSingleForm)
-                {
-                    index = -1;
-                }
-            }
-
-            return index;
-        }
-
-        /// <summary>
-        /// Creates an instance of child form of given form type
-        /// </summary>
-        /// <param name="formType">type of form to be created, for example SettingsXtraForm</param>
-        /// <returns>reference to newly created child form</returns>
-        private ChildForm CreateChildForm(Type formType, List<KeyValuePair<string, object>> parameters)
-        {
-            ChildForm newlyCreatedForm = null;
-
-            if (FramesCanOpen < 1)
-            {
-                HideProcessing();
-                this.ShowError(Program.LanguageManager.GetString(StringResources.Message_NoMoreDocumentsCanOpen),
-                    Program.LanguageManager.GetString(StringResources.Message_ErrorHeader));
-            }
-            else
-            {
-
-                newlyCreatedForm = (ChildForm)Program.Kernel.Get(formType,
-                        parameters.Select(kvp => new ConstructorArgument(kvp.Key, kvp.Value)).ToArray());
-
-                childForms[formType.Name].Add(newlyCreatedForm);
-                newlyCreatedForm.MdiParent = this;
-                newlyCreatedForm.FormClosed += ChildClosedEventHandler;
-                FramesCanOpen--;
-            }
-
-            return newlyCreatedForm;
-        }
-
-        /// <summary>
-        /// Cleans child form if it was closed
-        /// </summary>
-        /// <param name="sender">child form expected</param>
-        /// <param name="arguments"></param>
-        public void ChildClosedEventHandler(object sender, EventArgs arguments)
-        {
-            if(typeof(ChildForm).IsAssignableFrom(sender.GetType()))
-            {
-                foreach(var childType in childForms)
-                {
-                    if(childType.Value.Remove((ChildForm)sender))
-                    {
-                        FramesCanOpen++;
-                        break;
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Show existing child form.
-        /// </summary>
-        /// <param name="form"></param>
-        private void ShowChildForm(ChildForm form)
-        {
-            try
-            {
-                ShowProcessing();
-                form.Show();
-                form.WindowState = FormWindowState.Normal;
-                form.WindowState = FormWindowState.Maximized;
-            }
-            finally
-            {
-                HideProcessing();
-            }
-        }
-
-        /// <summary>
-        /// Create and show Settings child form. Starting tab page is set or first page if page doesn't exist.
-        /// </summary>
-        /// <param name="page">number of starting page</param>
-        public void CreateSettingsChildForm(int page)
-        {
-            try
-            {
-                SettingsXtraForm form = (SettingsXtraForm)OpenChildForm(typeof(SettingsXtraForm));
-
-                if (form != null && form.tabbedControlGroup.TabPages.Count > page)
-                {
-                    form.tabbedControlGroup.SelectedTabPage = form.tabbedControlGroup.TabPages[page];
-                }
-            }
-            finally
-            {
-                HideProcessing();
-            }
+            FormManager.Initialize(this, log);
         }
 
         /// <summary>
@@ -204,110 +83,17 @@ namespace Prizm.Main.Forms.MainChildForm
             {
                 parametrs.Add(new KeyValuePair<string, object>("number", number));
             }
-            return OpenChildForm(formType, parametrs, id);
+            return FormManager.Instance.OpenChildForm(formType, parametrs, id);
         }
 
         /// <summary>
-        /// Creation of child form. Can be used from outside to pass some Guid of entity and number to newly created forms.
+        /// Create and show Settings child form. Starting tab page is set or first page if page doesn't exist.
         /// </summary>
-        /// <param name="formType">exact type of form</param>
-        /// <param name="id">Guid of entity</param>
-        /// <param name="parameters">input parameters passed to the newly created form</param>
-        public ChildForm OpenChildForm(Type formType, List<KeyValuePair<string, object>> parameters, Guid id = default(Guid))
+        /// <param name="page">number of starting page</param>
+        public void CreateSettingsChildForm(int page)
         {
-
-            ChildForm form = null;
-            if (typeof(ChildForm).IsAssignableFrom(formType))
-            {
-                var forms = GetListChildForm(formType);
-                //search
-                int indexById = GetFormIndex(formType.Name, id);
-
-                if (indexById >= 0 && forms.Count > 0) // is open
-                {
-                    form = forms[indexById];
-                }
-                if (form != null)
-                {
-                    form.Activate();
-                }
-                else //create new
-                {
-                    int index = FindOpenForm(formType.Name, true);
-
-                    if (formType == typeof(RailcarNewEditXtraForm) && index >= 0)// this type form is open in edit mode
-                    {
-                        if (id == default(Guid))
-                        {
-                            string text = Program.LanguageManager.GetString(StringResources.MainWindow_CloseEditingReleaseNote);
-                            ShowWarning(text, "");
-                            form = forms[index];
-                            form.Activate();
-                        }
-                        else
-                        {
-                            string text = Program.LanguageManager.GetString(StringResources.MainWindow_OpenReleaseNoteReadOnly);
-                            if (this.ShowYesNo(text, ""))
-                            {
-                                ShowChildForm(formType, parameters, false);
-                            }
-                        }
-                    }
-                    else
-                    {
-                        ShowChildForm(formType, parameters, true);
-                    }
-                }
-            }
-            else
-            {
-                var e = new ApplicationException(String.Format("Could not create child form {0} because not of child type.", formType.Name));
-                log.Error(e.Message);
-                throw e;
-            }
-            return form;
+            FormManager.Instance.CreateSettingsChildForm(page);
         }
-
-        private void ShowChildForm(Type formType, List<KeyValuePair<string, object>> parameters, bool IsEditMode = true)
-        {
-            ChildForm form = CreateChildForm(formType, parameters);
-
-            if (form != null)
-            {
-                ShowChildForm(form);
-                if (!IsEditMode)
-                {
-                    ((ChildForm)form).IsEditMode = false;
-                }
-
-            }
-        }
-
-        /// <summary>
-        /// Find ChildForm index in corresponding childForms list that contains the element having mode equal ReadMode.
-        /// </summary>
-        /// <param name="formTypeName">string representation of form type name</param>
-        /// <param name="ReadMode"></param>
-        /// <returns>The zero-based index of the first occurrence of an element that matches the
-        /// conditions defined by match, if found; otherwise, –1.</returns>
-        private int FindOpenForm(string formTypeName, bool ReadMode)
-        {
-            int index = -1;
-            index = childForms[formTypeName]
-                .FindIndex(x => x.IsEditMode == ReadMode);
-            return index;
-        }
-
-        private List<ChildForm> GetListChildForm(Type formType)
-        {
-            if (!childForms.ContainsKey(formType.Name))
-            {
-                childForms.Add(formType.Name, new List<ChildForm>());
-            }
-            return childForms[formType.Name];
-        }
-
-
 
         #region Menu buttons
         private void barButtonItemNewPipe_ItemClick(object sender, ItemClickEventArgs e)
